requirements.txt
asgiref==3.2.7
boto3==1.13.0
Django==3.0.5
django-storages==1.9.1
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.1.2
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
django-cors-headers==3.2.1